refactor(session): separate shared daemon clients
