+++ src/session/client/capabilities.test.ts
