+++ src/session/client/capabilities.ts
