+++ src/session/client/daemonHttp.test.ts
